# frontend/openapi.json
{"openapi":"3.0.0","paths":{"/api/v1/items":{"post":{"operationId":"ItemsController_create_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateItemDto"}}}},"responses":{"201":{"description":"The item has been successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ItemResponseDto"}}}},"401":{"description":"Unauthorized."}},"security":[{"bearer":[]}],"summary":"Create a new item","tags":["items"]},"get":{"operationId":"ItemsController_findAll_v1","parameters":[{"name":"page","required":false,"in":"query","description":"Page number","schema":{"default":1,"type":"number"}},{"name":"limit","required":false,"in":"query","description":"Number of results per page","schema":{"default":10,"type":"number"}}],"responses":{"200":{"description":"Items retrieved successfully.","content":{"application/json":{"schema":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/ItemResponseDto"}},"count":{"type":"number"}}}}}}},"summary":"Get all items with pagination","tags":["items"]}},"/api/v1/items/{id}":{"get":{"operationId":"ItemsController_findById_v1","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Item retrieved successfully.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ItemResponseDto"}}}},"404":{"description":"Item not found."}},"summary":"Get an item by id","tags":["items"]},"patch":{"operationId":"ItemsController_update_v1","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateItemDto"}}}},"responses":{"200":{"description":"The item has been successfully updated.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ItemResponseDto"}}}},"401":{"description":"Unauthorized."},"403":{"description":"Forbidden."},"404":{"description":"Item not found."}},"security":[{"bearer":[]}],"summary":"Update an item","tags":["items"]},"delete":{"operationId":"ItemsController_remove_v1","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The item has been successfully deleted."},"401":{"description":"Unauthorized."},"403":{"description":"Forbidden."},"404":{"description":"Item not found."}},"security":[{"bearer":[]}],"summary":"Delete an item","tags":["items"]}},"/api/v1/auth/me":{"get":{"operationId":"AuthController_me_v1","parameters":[],"responses":{"200":{"description":"User successfully retrieved.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}},"401":{"description":"Invalid credentials."}},"security":[{"bearer":[]}],"summary":"Get current user","tags":["auth"]}},"/api/v1/auth/register":{"post":{"operationId":"AuthController_register_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"User successfully registered.","content":{"application/json":{"schema":{"type":"object","properties":{"accessToken":{"type":"string"},"user":{"type":"object"}}}}}},"409":{"description":"Email or username already exists."}},"summary":"Register a new user","tags":["auth"]}},"/api/v1/auth/login":{"post":{"operationId":"AuthController_login_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"User successfully logged in, returns access token and sets refresh token in HttpOnly cookie.","content":{"application/json":{"schema":{"type":"object","properties":{"accessToken":{"type":"string"}}}}}},"401":{"description":"Invalid credentials."}},"summary":"Log in a user","tags":["auth"]}},"/api/v1/auth/logout":{"post":{"operationId":"AuthController_logout_v1","parameters":[],"responses":{"200":{"description":"User successfully logged out.","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"}}}}}}},"security":[{"bearer":[]}],"summary":"Log out a user","tags":["auth"]}},"/api/v1/auth/refresh":{"post":{"operationId":"AuthController_refreshTokens_v1","parameters":[],"responses":{"200":{"description":"Tokens successfully refreshed.","content":{"application/json":{"schema":{"type":"object","properties":{"accessToken":{"type":"string"}}}}}},"401":{"description":"Invalid refresh token."}},"summary":"Refresh access token","tags":["auth"]}}},"info":{"title":"CRUD Platform API","description":"API for managing users and contents in a CRUD platform.","version":"1","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateItemDto":{"type":"object","properties":{"title":{"type":"string","example":"My First Item","description":"Item title"},"content":{"type":"string","example":"This is the content of my first item.","description":"Item content"}},"required":["title","content"]},"AuthorResponseDto":{"type":"object","properties":{"id":{"type":"string"},"username":{"type":"string"}},"required":["id","username"]},"ItemResponseDto":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"content":{"type":"string"},"author":{"$ref":"#/components/schemas/AuthorResponseDto"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","title","content","author","createdAt","updatedAt"]},"UpdateItemDto":{"type":"object","properties":{"title":{"type":"string","example":"My First Item","description":"Item title"},"content":{"type":"string","example":"This is the content of my first item.","description":"Item content"}}},"UserResponseDto":{"type":"object","properties":{"id":{"type":"string"},"email":{"type":"string"},"username":{"type":"string"}},"required":["id","email","username"]},"CreateUserDto":{"type":"object","properties":{"username":{"type":"string","example":"john_doe","description":"Username"},"email":{"type":"string","example":"user@example.com","description":"User email"},"password":{"type":"string","example":"password123","description":"User password"}},"required":["username","email","password"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","example":"user@example.com","description":"User email"},"password":{"type":"string","example":"password123","description":"User password"}},"required":["email","password"]}}}}